simulate_sample_path rename
Renamed hot_deck_forecast to this, b/c the actual forecast call will need to generate multiple paths, and this function generates a single path. So, it should also be a non-exported function. Also, renamed n to n_closest, b/c need another `n` in the to be created exported fn for the # of paths to generate.