train-test split bug fix
The train test split's process of moving the `mobile_data` to before the `ante_split` left a gap b/w the two, resulting in the final row in `mobile_data` having the first row of `ante_split` as.. its next row.

This led to using that next row as the "lead" (i.e. next) value in the hot deck process, whereas it should have been NA.

As such, I applied `fill_gaps` to put all of the dates b/w these two rows' dates in between them, so that the final row of `mobile_data` has an NA value as its next observation.