get_local_rows fully index-based
This operates at ~2-3x the speed.

The code is a good bit harder to logic through, though so I've retained the date-based implementation in the file "slow_hotdeckfc.R", and prepended the relevant methods with "slow_". I had to modify the slow_get_local_rows slightly to match the new implementation's handling of leap years. I used this old implementation to test against the new, using a real dataset (SUGG).

I tried to annotate the new code as well as I could.