train-test split max(window_fwd)
At least it was a simple fix!

I am using window_fwd in the train-test split, so when I allowed window_fwd to be a vector, I needed to tell tt-split what to do in that case. I'll just take the max, because (a) it's supposed to be conservative, and (b) that data shouldn't be near any window anyways.