mutators
These are intended to be passed to CV, so that desired mutations that would cause data leakage in the train-test split can be applied after said split. For instance, if you want a column of leads (next observations), it should be NA for the final training observation before the first test observation.