incorporated sampler design
This is the major redesign that allows for arbitrary functions ("sampler"s) to be used to generate simulated sample values, given the "current observation" and local observations. After receiving the first "current observation", a custom sampler could return any object it wants as it's next-step (next h) new current observation, thanks to R's typing laxness. For instance, you could return a distribution or a vector.

I still need to settle on how I want to build the grid for grid searching.