diff sampler implemented
Instead of taking the next observation of the sampled row, it takes the difference from the current to next observation of said sampled row, then adds that to the current observation.

The idea behind this is that it may help avoid getting trapped in cycles at the bounds of the response variable.