memoized get_local_rows
This *appears* correct. I tested it by inspection with the SUGG and CRAM data on a set seed in CV. I think I *could* set up a test for this, but it'd be a lot of work, a time-consuming test, and amount to repeatedly re-doing said test-by-inspection.

As it is, this memoization looks like it cuts the run time of CV down to a quarter of what it was.

refs: https://memoise.r-lib.org/reference/memoise.html#details
https://rdrr.io/r/base/ns-hooks.html
https://github.com/rstudio/profvis/issues/134
https://github.com/r-lib/memoise/issues/79#issuecomment-961492685