added layer to hot_deck_forecast
Abstracting it to be able to take covariate forecasts. The premise is to have hot_deck_forecast act largely as a switch: if there's no covariates, just call the internal method, which is just the *old* hot_deck_forecast (with the validation moved to this new h_d_f). If there are covariates, then slice them up by their simulation number, put them in the sampler, and iterate on the old h_d_f.